+++ cx-portal/src/components/pages/AppReleaseProcess/components/AppMarketCard/index.tsx
@@ -32,6 +32,7 @@ import {
   SelectList,
   UploadFileStatus,
   PageSnackbar,
+  UploadStatus,
 } from 'cx-portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import { Grid, Divider, Box, InputLabel } from '@mui/material'
@@ -100,6 +101,7 @@ export const ConnectorFormInputField = ({
   maxFilesToUpload,
   maxFileSize,
   defaultValues,
+  enableDeleteIcon,
 }: any) => (
   <Controller
     name={name}
@@ -139,6 +141,7 @@ export const ConnectorFormInputField = ({
             acceptFormat={acceptFormat}
             maxFilesToUpload={maxFilesToUpload}
             maxFileSize={maxFileSize}
+            enableDeleteIcon={enableDeleteIcon}
           />
         )
       } else if (type === 'checkbox') {
@@ -236,7 +239,6 @@ export default function AppMarketCard() {
     price: appStatusData?.price,
     useCaseCategory: appStatusData?.useCase,
     appLanguage: appStatusData?.supportedLanguageCodes,
-    //To do: to be changed once api is available
     salesManagerId: appStatusData?.salesManagerId,
     shortDescriptionEN:
       appStatusData?.descriptions?.filter(
@@ -441,14 +443,14 @@ export default function AppMarketCard() {
                 status,
               } as any)
 
-            setFileStatus('uploading')
+            setFileStatus(UploadStatus.UPLOADING)
 
             uploadDocumentApi(result, 'APP_LEADIMAGE', uploadImageValue)
               .then(() => {
-                setFileStatus('upload_success')
+                setFileStatus(UploadStatus.UPLOAD_SUCCESS)
               })
               .catch(() => {
-                setFileStatus('upload_error')
+                setFileStatus(UploadStatus.UPLOAD_ERROR)
               })
 
             dispatch(setAppId(result))