+++ cx-portal/src/components/pages/Test/index.newdropzone.tsx
@@ -20,7 +20,12 @@
 
 import Box from '@mui/material/Box'
 import { Dropzone, DropzoneFile } from 'components/shared/basic/Dropzone'
-import { Button, DropArea, UploadFileStatus } from 'cx-portal-shared-components'
+import {
+  Button,
+  DropArea,
+  UploadFileStatus,
+  UploadStatus,
+} from 'cx-portal-shared-components'
 import React from 'react'
 import { Controller, useForm } from 'react-hook-form'
 
@@ -48,14 +53,14 @@ export default function NewDropzoneTest() {
     }
 
     for (let fileIndex = 0; fileIndex < myUploadFiles.length; fileIndex++) {
-      setFileStatus(fileIndex, 'uploading')
+      setFileStatus(fileIndex, UploadStatus.UPLOADING)
 
       try {
         await fetch('https://httpbin.org/post', { method: 'POST' })
-        setFileStatus(fileIndex, 'upload_success')
+        setFileStatus(fileIndex, UploadStatus.UPLOAD_SUCCESS)
       } catch (e) {
         // TODO: it should display an error notification snackbar
-        setFileStatus(fileIndex, 'upload_error')
+        setFileStatus(fileIndex, UploadStatus.UPLOAD_ERROR)
       }
     }
   }